requirements.txt
Flask==3.1.0
numpy==2.2.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.1.0
pymongo==4.10.1
python-dotenv==1.0.1
python_bcrypt==0.3.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
Werkzeug==3.1.3
